Make git_url parameter useful
Without this patch, the source repository for the bugdaystats vcsrepo
is hard-coded and the $bugdaystats::site::git_url parameter is unused.
This patch adds the flexibility that was presumably intended by that
parameter originally.

An alternative to this patch is to remove the parameter completely,
since users are not likely to use this module with another repo. This
would require a patch to system-config and would break other users
currently trying to use this parameter (even though it doesn't do
anything).

Change-Id: If5e547c9270a05359a1bacbfe5dfbe4ec28ed383